Really bump to 10.0.2 now (#16)
Co-authored-by: Florian Magin <fmagin@users.noreply.github.com>

turns out that I messed up last time and only changed the version string not the direct link to the zip which was the actually important part. Now the CI just uses the setup-ghidra action with just the version number again which has been fixed and works with the github releases by default now.